pyproject.toml
[project]
name = "taboo-words"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "openai>=1.37.1",
    "pandas>=2.2.2",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"taboo-words" = "taboo_words:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/taboo_words"]
